Revert change made to bilinear sampler in libmv
This made preview working but that broke internals
of tracking.

Namely, BlurredImageAndDerivativesChannels is giving
much more blurred image because it was assuming pixel
center is an integer position.

Guess other parts of libmv used to suffer because of
this issue.

Now pixel centering happens in blender side, and
libmv assumes integer position is a pixel center.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55844 954f8c5b-7b00-dc11-b283-0030488c597c