Fix #35081: opening .blend files with chinese characters not working. For compressed files we were not passing the full wide char path to zlib, so not all file names worked. Now we use gzopen_w available in new zlib versions.
Patch by Tamito Kajiyama, I added an extra check for the zlib version so it
keeps compiling with older versions for now.


For platform maintainers:

Part of this commit are zlib 1.2.8 libraries for windows 32 bit. We still
need update libraries for windows 64 bit and mingw. There's a readme.txt
and build.bat included with instructions on how to build.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@56454 954f8c5b-7b00-dc11-b283-0030488c597c