NLA Editor: It is now possible to add strips to AnimData blocks with no existing tracks
As a convenience feature for those who are loading in action libraries and using
these to quickly block out things in the NLA editor, it is now possible to add
strips to AnimData blocks without first manually creating empty tracks to add
these strips to. Simply ensure that such empty AnimData blocks are selected
(Hint: click on the action line of the affected AnimData block to do so), and
try to add a strip normally.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@57312 954f8c5b-7b00-dc11-b283-0030488c597c